+++ src/components/firms.jl
@@ -4,51 +4,51 @@ abstract type FirmComponent <: AbstractComponent end
 end
 
 @component struct LaborProductivity <: FirmComponent #alpha_bar_i
-    value::Float64
+    value::FloatType
 end
 
 @component struct IntermediateProductivity <: FirmComponent #beta_i
-    value::Float64
+    value::FloatType
 end
 
 @component struct CapitalProductivity <: FirmComponent #kappa_i
-    value::Float64
+    value::FloatType
 end
 
 @component struct CapitalDeprecationRate <: FirmComponent #delta_i
-    rate::Float64
+    rate::FloatType
 end
 
 @component struct OperatingMargins <: FirmComponent  #pi_bar_i
-    rate::Float64
+    rate::FloatType
 end
 
 @component struct WageBill <: FirmComponent #w_i
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct AverageWageRate <: FirmComponent #w_bar_i
-    rate::Float64
+    rate::FloatType
 end
 
 @component struct OutputTaxRate <: FirmComponent #tau_Y
-    rate::Float64
+    rate::FloatType
 end
 
 @component struct CapitalTaxRate <: FirmComponent #tau_k
-    rate::Float64
+    rate::FloatType
 end
 
 @component struct Price <: FirmComponent #P_i
-    value::Float64
+    value::FloatType
 end
 
 @component struct PriceIndex <: FirmComponent #P_bar_i
-    value::Float64
+    value::FloatType
 end
 
 @component struct CFPriceIndex <: FirmComponent #P_CF_i
-    value::Float64
+    value::FloatType
 end
 
 @component struct Employment <: FirmComponent #N_i
@@ -64,35 +64,35 @@ end
 end
 
 @component struct Output <: FirmComponent #Y_i
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct Sales <: FirmComponent #Q_i
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct GoodsDemand <: FirmComponent #Q_d_i
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct Inventories <: FirmComponent #S_i
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct Intermediates <: FirmComponent #M_i
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct Investment <: FirmComponent  #I_i
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct Equity <: FirmComponent #E_i
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct FinalGoodsStockChange <: FirmComponent #DS_i
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct MaterialsStockChange <: FirmComponent #DM_i
