+++ src/components/government.jl
@@ -1,27 +1,27 @@
 abstract type GovernmentComponent <: AbstractComponent end
 
 @component struct GovernmentRevenues <: GovernmentComponent #Y_G
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct SocialBenefitsInactive <: GovernmentComponent #sb_inact
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct SocialBenefitsOther <: GovernmentComponent #sb_other
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct PriceInflationGovernmentGoods <: GovernmentComponent #P_j
-    value::Float64
+    value::FloatType
 end
 
 @component struct GovernmentDebt <: GovernmentComponent #L_G
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct ConsumptionDemand <: GovernmentComponent #C_G
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct LocalGovernment <: Ark.Relationship end
