+++ src/components/households.jl
@@ -15,18 +15,18 @@ end
 end
 
 @component struct ConsumptionBudget <: AbstractComponent
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct InvestmentBudget <: AbstractComponent
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct RealisedConsumption <: AbstractComponent
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct RealisedInvestment <: AbstractComponent
-    amount::Float64
+    amount::FloatType
 end
 @component struct Household <: AbstractComponent end
