+++ src/components/profits.jl
@@ -1,7 +1,7 @@
 @component struct Profits <: AbstractComponent
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct ExpectedProfits <: AbstractComponent
-    amount::Float64
+    amount::FloatType
 end
