+++ src/components/rest_of_world.jl
@@ -1,58 +1,58 @@
 @component struct EuroAreaGDP <: AbstractComponent
-    value::Float64
+    value::FloatType
 end
 
 @component struct EuroAreaGrowth <: AbstractComponent
-    rate::Float64
+    rate::FloatType
 end
 
 @component struct EuroAreaInflation <: AbstractComponent
-    rate::Float64
+    rate::FloatType
 end
 
 @component struct NetForeignPosition <: AbstractComponent
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct ImportSupply <: AbstractComponent
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct TotalExportDemand <: AbstractComponent
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct TotalImportSupply <: AbstractComponent
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct ExportDemand <: AbstractComponent
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct ImportSales <: AbstractComponent
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct ImportDemand <: AbstractComponent
-    amount::Float64
+    amount::FloatType
 end
 @component struct ImportPrice <: AbstractComponent
-    value::Float64
+    value::FloatType
 end
 
 @component struct ExportPriceInflation <: AbstractComponent
-    value::Float64
+    value::FloatType
 end
 
 @component struct ForeignSector <: AbstractComponent end
 
 @component struct ForeignConsumptionDemand <: AbstractComponent
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct ForeignConsumption <: AbstractComponent
-    amount::Float64
+    amount::FloatType
 end
 
 @component struct RestOfWorldEntity <: AbstractComponent
